+++ packages/client/src/widgets/Game/ui/Game/models/collisionUtils.ts
